% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{accuracy}
\alias{accuracy}
\title{Compute classification accuracy.}
\usage{
accuracy(estimate, class)
}
\arguments{
\item{estimate}{The estimated class}

\item{class}{The true class}
}
\value{
The classification accuracy
}
\description{
Compute classification accuracy.
}

