% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{class_probability_ranking_randomforest}
\alias{class_probability_ranking_randomforest}
\title{Using the randomForest classifier, calculate the Spearman
rank-order correlation of the probability of belonging to class 0
(suitable only for a binary class dataset) between the original
dataset and a dataset permuted as parametrized by the tree.}
\usage{
class_probability_ranking_randomforest(model, tree, data, class = "PClass")
}
\arguments{
\item{model}{A model of the data from a classifier}

\item{tree}{A permutation tree}

\item{data}{The data matrix}

\item{class}{The name of the column with the predicted class label}
}
\value{
Correlation between original and permuted class probabilities
}
\description{
Using the randomForest classifier, calculate the Spearman
rank-order correlation of the probability of belonging to class 0
(suitable only for a binary class dataset) between the original
dataset and a dataset permuted as parametrized by the tree.
}

