% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artificial_data.R
\name{classifier.double.xor}
\alias{classifier.double.xor}
\title{Exact classifier for the Double-XOR data}
\usage{
classifier.double.xor(formula, data)
}
\arguments{
\item{formula}{Model formula}

\item{data}{The data}
}
\description{
Exact classifier for the Double-XOR data
}

