% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gen.data.single.xor}
\alias{gen.data.single.xor}
\title{Generate random binary data (Single-XOR)}
\usage{
gen.data.single.xor(nrows = 1000, ncols = 5, noiselevel = 0.1,
  noise = TRUE)
}
\arguments{
\item{nrows}{Number of rows in the data}

\item{ncols}{Number of columns in the data}

\item{noiselevel}{Fraction of class labels to randomly flip.}

\item{noise}{Add noise or not (Boolean).}
}
\value{
Data matrix where Class = XOR(1, 2) | 3
}
\description{
Generate random binary data (Single-XOR)
}

