% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldeneye-package.R
\docType{package}
\name{goldeneye-package}
\alias{goldeneye}
\alias{goldeneye-package}
\title{goldeneye: Sharing Secrets Safely}
\description{
The goldeneye package is designed to allow data to safely be shared between R users over insecure channels. While it does require using R, the interfaces are written to be as simple as possible in order to reduce the technical barrier for beginner R users. This package was developed as part of the NordForsk-funded DigiVet project.
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

Other contributors:
\itemize{
  \item Mossa Merhi Reimert \email{mossa@sund.ku.dk} [contributor]
  \item Maya Katrin Gussmann \email{makg@sund.ku.dk} [contributor]
  \item Søren Saxmose Nielsen \email{saxmose@sund.ku.dk} [contributor]
  \item Francisco Fernando Calvo-Artavia \email{FRAFE@fvst.dk} [contributor]
  \item Ivana Rodriguez Ewerlöf \email{ivana.ewerlof@sva.se} [contributor]
  \item Jessica Enright \email{Jessica.Enright@glasgow.ac.uk} [contributor]
}

}
\keyword{internal}
