% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{group.importance}
\alias{group.importance}
\title{Determine the relative importance of each group in the grouping}
\usage{
group.importance(model = model, data = data, class.name = class.name,
  class.index = class.index, tree = tree)
}
\arguments{
\item{model}{A model of the data from a classifier}

\item{data}{The data matrix}

\item{class.name}{The name of the column with the class label}

\item{class.index}{The name of the column with the class label}

\item{tree}{A permutation tree}
}
\value{
A vector with the relative fidelity of each group.
}
\description{
Determine the relative importance of each group in the grouping
}
\keyword{internal}

