% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gy_join_group.R
\name{gy_join_group}
\alias{gy_join_group}
\alias{gy_set_group}
\title{Join an existing goldeneye user group via a weblink}
\usage{
gy_join_group(weblink = NULL, user = NULL, make_default = FALSE)

gy_set_group(group, make_default = FALSE, silent = FALSE)
}
\arguments{
\item{weblink}{the weblink as provided by the group administrator}

\item{user}{the username that you want to use for this group}
}
\description{
Join an existing goldeneye user group via a weblink
}
