% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gy_key_funs.R
\name{gy_key_funs}
\alias{gy_key_funs}
\title{Obtain key functions}
\usage{
gy_key_funs(type)
}
\arguments{
\item{type}{type of key function (only "identity" is currently supported)}
}
\description{
Obtain key functions
}
