% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gy_check.R
\name{gy_profile}
\alias{gy_profile}
\alias{gy_info}
\title{Set a goldeneye profile and display user information}
\usage{
gy_profile(path = getOption("goldeneye_path"), silent = FALSE)

gy_info()
}
\arguments{
\item{path}{path to goldeneye profile}

\item{silent}{option to suppress output to screen}
}
\description{
Set a goldeneye profile and display user information
}
