% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gy_public_file.R
\name{gy_public_file}
\alias{gy_public_file}
\title{Extract the public key file}
\usage{
gy_public_file(file = "goldeneye_public.gyu", silent = FALSE)
}
\arguments{
\item{file}{filename to write to}

\item{silent}{do this quietly?}
}
\description{
Extract the public key file
}
