% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gy_saveRDS.R
\name{gf_saveRDS}
\alias{gf_saveRDS}
\alias{gy_saveRDS}
\alias{gy_readRDS}
\title{Save and read encrypted RDS}
\usage{
gy_saveRDS(
  object,
  file = stop("file must be specified (.rdg file extension is recommended)"),
  users = character(0),
  local_user = TRUE,
  comment = "",
  overwrite = FALSE,
  ascii = FALSE,
  funs = list(type = "identity"),
  ...
)

gy_readRDS(
  file = stop("file must be specified (.rdg file extension is recommended)"),
  run_custom = TRUE
)
}
\arguments{
\item{object}{the object to encrypt}

\item{file}{the filename to which the encrypted object will be saved}

\item{users}{one or more users who will be authorised to decrypt the file}

\item{local_user}{should the local user be able to decrypt the file?}

\item{comment}{a comment to include}

\item{overwrite}{should the file be overwritten, if it exists?}

\item{ascii}{argument passed to saveRDS}

\item{funs}{argument passed to saveRDS}

\item{...}{arguments passed to gy_serialise}

\item{run_custom}{should any custom decryption functions be run automatically?}
}
\description{
Save and read encrypted RDS
}
