% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gy_zip.R
\name{gy_zip}
\alias{gy_zip}
\alias{gy_unzip}
\title{Zip and encrypt external files using goldeneye}
\usage{
gy_zip(
  input_files,
  file = stop("file must be specified (.rfg file extension is recommended)"),
  compression_level = 9,
  users = character(0),
  local_user = TRUE,
  comment = "",
  funs = gy_key_funs("identity")
)

gy_unzip(
  file = stop("file must be specified (.rfg file extension is recommended)"),
  directory = getwd(),
  overwrite = FALSE,
  run_custom = TRUE
)
}
\arguments{
\item{compression_level}{compression level to use on a scale of 1-9 (passed to \code{\link[zip]{zip}})}

\item{users}{a character vector of (other) users within your current user group for whom the encrypted file will be decryptable. Alternatively, this can be a vector of paths/urls to public keys, or a mixture of the two.}

\item{local_user}{should the current user also be able to decrypt thw file?}

\item{comment}{an optional comment that will be sent (unencrypted) along with the file}

\item{funs}{optional additional encryption steps: this must be the output of a call to \code{\link{gy_key_funs}}}

\item{files}{one or more files to add to the zip archive}
}
\description{
Zip and encrypt external files using goldeneye
}
