\name{permutedata}
\alias{permutedata}
\title{Use the permutation to obtain a new data matrix.}
\usage{
permutedata(data, perm)
}
\arguments{
  \item{tree}{The permutation tree}

  \item{perm}{The permutation}
}
\value{
The permuted data matrix
}
\description{
Use the permutation to obtain a new data matrix.
}
\keyword{internal}

