\name{predict.classifier.single.xor}
\alias{predict.classifier.single.xor}
\title{Prediction function for the Single-XOR}
\usage{
\method{predict}{classifier.single.xor}(object, newdata, ...)
}
\arguments{
  \item{object}{Object}

  \item{newdata}{The data to predict}

  \item{...}{Additional arguments}
}
\description{
Prediction function for the Single-XOR
}

