% Generated by roxygen2 (4.0.2): do not edit by hand
\name{accuracy}
\alias{accuracy}
\title{Compute classification accuracy.}
\usage{
accuracy(estimate, class)
}
\arguments{
\item{estimate}{The estimated class}

\item{class}{The true class}
}
\value{
The classification accuracy
}
\description{
Compute classification accuracy.
}
\keyword{internal}

