\name{add.data.noise}
\alias{add.data.noise}
\title{Flip the class label in noiselevel fraction of the data}
\usage{
add.data.noise(classvector, noiselevel)
}
\arguments{
  \item{classvector}{A vector giving the class of the data}

  \item{noiselevel}{Fraction of class labels to randomly
  flip.}
}
\value{
Class vector with randomly noiselevel amount of random
noise.
}
\description{
Flip the class label in noiselevel fraction of the data
}
\keyword{internal}

