% Generated by roxygen2 (4.0.2): do not edit by hand
\name{classifier.single.xor}
\alias{classifier.single.xor}
\title{Exact classifier for the Single-XOR data}
\usage{
classifier.single.xor(formula, data)
}
\arguments{
\item{formula}{Model formula}

\item{data}{The data}
}
\description{
Exact classifier for the Single-XOR data
}

