% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gy_encrypt.R
\name{gy_encrypt}
\alias{gy_encrypt}
\alias{gy_decrypt}
\alias{get_public_keys}
\title{Encrypt and decrypt a pre-serialised object using goldeneye}
\usage{
gy_encrypt(
  object,
  users = character(0),
  local_user = TRUE,
  comment = "",
  funs = gy_key_funs("identity")
)

gy_decrypt(object, run_custom = TRUE)

get_public_keys(users, type = "curve")
}
\arguments{
\item{object}{the serialised object to encrypt}

\item{users}{a character vector of (other) users within your current user group for whom the encrypted file will be decryptable. Alternatively, this can be a vector of paths/urls to public keys, or a mixture of the two.}

\item{local_user}{should the current user also be able to decrypt thw file?}

\item{comment}{an optional comment that will be sent (unencrypted) along with the file}

\item{funs}{optional additional encryption steps: this must be the output of a call to \code{\link{gy_key_funs}}}

\item{run_custom}{should any custom decryption functions be run automatically?}

\item{type}{the type of public key to return ("curve" or "ed")}
}
\description{
Encrypt and decrypt a pre-serialised object using goldeneye
}
