% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gy_serialise.R
\name{gy_serialise}
\alias{gy_serialise}
\alias{gy_deserialise}
\alias{gy_serialize}
\alias{gy_deserialize}
\title{Serialise and deserialise a list of objects/files}
\usage{
gy_serialise(object, files = character(0), method = "qs", ...)

gy_deserialise(object, files = TRUE, ...)

gy_serialize(object, files = character(0), method = "qs", ...)

gy_deserialize(object, files = TRUE, ...)
}
\arguments{
\item{object}{the object to encrypt}

\item{files}{external files to serialise (not yet implemented)}

\item{method}{the serialisation method to use (base, qs, or custom)}

\item{...}{arguments passed to the underlying serialisation method}
}
\description{
Serialise and deserialise a list of objects/files
}
