% Generated by roxygen2 (4.0.2): do not edit by hand
\name{permutation}
\alias{permutation}
\title{Permute the data matrix using the permutation tree.}
\usage{
permutation(tree, data)
}
\arguments{
\item{tree}{The permutation tree}

\item{data}{the data matrix}
}
\value{
A permutation of the indices of the rows in the data matrix
}
\description{
Permute the data matrix using the permutation tree.
}

