% Generated by roxygen2 (4.0.2): do not edit by hand
\name{permutedata}
\alias{permutedata}
\title{Use the permutation to obtain a new data matrix.}
\usage{
permutedata(data, perm)
}
\arguments{
\item{perm}{The permutation}

\item{tree}{The permutation tree}
}
\value{
The permuted data matrix
}
\description{
Use the permutation to obtain a new data matrix.
}
\keyword{internal}

