% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.classifier.double.xor}
\alias{predict.classifier.double.xor}
\title{Prediction function for the Double-XOR data}
\usage{
\method{predict}{classifier.double.xor}(object, newdata, ...)
}
\arguments{
\item{object}{Object}

\item{newdata}{The data to predict}

\item{...}{Additional arguments}
}
\description{
Prediction function for the Double-XOR data
}

