% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict_randomforest_probability}
\alias{predict_randomforest_probability}
\title{A predict function for the randomForest classifier, which returns a
vector giving the probability of instances belonging to class 0.}
\usage{
predict_randomforest_probability(model, newdata)
}
\arguments{
\item{model}{A model of the data from a classifier}

\item{newdata}{}
}
\description{
A predict function for the randomForest classifier, which returns a
vector giving the probability of instances belonging to class 0.
}

